Fixed an issue with copying the config.
On some versions of python with an up to date setuptools it will
create an armory*.egg for the package instead of creating a directory
to install within the {dist,site}-packages. Therefore, changed it to
use pkg_resources.resource_string to get the file contents and write it
to the home directory location. Included a __init__.py within the armory.default_configs
directory as well to make it a module.

Also changed the default sqlite database name to include armory and the
default "path/directory" output goes to the current directory.